refactor(frontend): move mailbox events to shared global store
